experimental GSoC 2018 Python 3 support by Sanjeet Bhatti
git-svn-id: https://svn.osgeo.org/grass/grass/trunk@73229 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ lib/python/pygrass/vector/geometry.py
@@ -721,7 +721,8 @@ class Line(Geo):
                                           pnt.c_points.contents.x,
                                           pnt.c_points.contents.y,
                                           pnt.c_points.contents.z,
-                                          angle, slope):
+                                          ctypes.c_double(angle),
+                                          ctypes.c_double(slope)):
             raise ValueError("Vect_point_on_line give an error.")
         pnt.is2D = self.is2D
         return pnt
